undo last commit

+++ gflib/sprite.h
@@ -29,10 +29,10 @@ struct SpriteFrameImage
     bool8 relativeFrames;
 };
 
-#define obj_frame_tiles(ptr) {.data = (u16 *)ptr, .size = sizeof ptr}
+#define obj_frame_tiles(ptr) {.data = (u8 *)ptr, .size = sizeof ptr}
 
-#define overworld_frame(ptr, width, height, frame) {.data = (u16 *)ptr + (width * height * frame * 64)/2, .size = (width * height * 64)/2}
-#define overworld_ascending_frames(ptr, width, height) {.data = (u16 *)ptr, .size = (width * height * 64)/2, .relativeFrames=TRUE}
+#define overworld_frame(ptr, width, height, frame) {.data = (u8 *)ptr + (width * height * frame * 64)/2, .size = (width * height * 64)/2}
+#define overworld_ascending_frames(ptr, width, height) {.data = (u8 *)ptr, .size = (width * height * 64)/2, .relativeFrames=TRUE}
 
 struct SpritePalette
 {
